Build libsmi 0.5.0 into the base image.
The CentOS repositories do not provide a version of the libsmi package newer than version 0.4.8.  The 0.4.8 version has a bug that causes infinite loop and memory exhaustion when parsing a MIB node definition containing overlapping range definitions.  Version 0.5.0 of libsmi fixes that issue.

This change builds an RPM out of the libsmi 0.5.0 source archive, installs it, and makes the zenoss-centos-deps package depend on it.

Fixes ZEN-30018.